added script that creates the MySQL database
this .sql file includes the script that creates the university_sync database with essentially needed tables in it. This script declares the columns, data types they are storing, primary keys and foreign keys. You need to run this file once (after of course you have MySQL version 8 server and MySQL workbench installed) in order to have the database and be able to do operations on it. Beware that script first deletes the database if it exists (along with all data in it), and then recreates it from scratch. Therefore it should only be run once, or as a "recovery" option whenever an irreversible harmful change is unintentionally done on the database.