started DAO implementation
added a DBConnectionManager that creates the connection. Added UserDAO class that can insert lines into User table based on the input given to its method. Added a Main file to be able to test this feature. Note that this is the very beginning of DAO implementation and also we need to declare packages as it should be in the main project. This DAO also assumes that you have a valid table initialized by the .sql script in this folder and also a valid connection.